+++ packages/tinacms/src/toolkit/form-builder/index.ts
@@ -8,4 +8,5 @@ export * from './loading-dots';
 export * from './reset-form';
 export * from './form-actions';
 export * from './create-branch-modal';
+export * from './EmbeddedChatbot';
 export type { FieldRenderProps } from 'react-final-form';