+++ packages/next-tinacms-s3/CHANGELOG.md
@@ -1,5 +1,12 @@
 # next-tinacms-s3
 
+## 17.0.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - tinacms@3.3.0
+
 ## 16.0.0
 
 ### Patch Changes